% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rfm_rdatetime}
\alias{rfm_rdatetime}
\title{Generate random dateteime sequence}
\usage{
rfm_rdatetime(n, begin = as.POSIXct("2014-12-01", tz = Sys.timezone()),
  end = as.POSIXct("2014-12-31", tz = Sys.timezone()))
}
\description{
Generate random dateteime sequence
}

