% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rfm_generate_data}
\alias{rfm_generate_data}
\title{Generate sample data for RFM analysis}
\usage{
rfm_generate_data(id_num = 1000, date_type = c("char", "Date", "POSIXct",
  "POSIXlt"), seed, ...)
}
\description{
Generate sample data for RFM analysis
}

